only load redis when running in lambda
- had to move around resolveser to make sure that we dont load redis through imports
(blocks lambda from responding until request fails)
