+++ src/api/api.ts
@@ -12,10 +12,11 @@ import crypto from 'crypto'
 // used to encrypt request ip
 const secret = process.env.HASH_SECRET || Math.random().toString()
 
-import { typeDefs, resolvers } from './graphql'
+import { typeDefs, vendors, bestPrices } from './graphql'
 import { SOURCES, RELOAD_TYPES, AMMO_TYPES, TYPES } from '../constants'
 import { ItemType } from '../graphql-types'
-
+import { getItemsFlatListings, getScrapeResponses } from './shared'
+import { getRedisItems } from './redis-getter'
 const DEV = process.env.DEV === 'true'
 
 const client = redis.createClient({ host: 'redis' })
@@ -43,7 +44,7 @@ server.route({
 server.route({
   method: 'POST',
   path: '/track-performance',
-  handler: function (request, h) {
+  handler: function(request, h) {
     const userAgent = request.headers['user-agent'] || 'unknown'
     const { performance, href } =
       typeof request.payload === 'string'
@@ -69,7 +70,7 @@ server.route({
 server.route({
   method: 'POST',
   path: '/track-view',
-  handler: function (request, h) {
+  handler: function(request, h) {
     // record user agent + calibre + brand that user opened up
     const userAgent = request.headers['user-agent'] || 'unknown'
     const body =
@@ -111,7 +112,7 @@ server.route({
 server.route({
   method: 'POST',
   path: '/track-click',
-  handler: async function (request, h) {
+  handler: async function(request, h) {
     // record user agent + calibre + brand that user opened up
     const body =
       typeof request.payload === 'string'
@@ -185,7 +186,7 @@ server.route({
     },
   },
   path: '/content-security-report-uri',
-  handler: function (req, h) {
+  handler: function(req, h) {
     let body = {}
     try {
       body =
@@ -244,7 +245,7 @@ server.events.on('response', (request: Request) => {
       },
       string: `Error response (500) sent for request: ${request.method.toUpperCase()} ${
         request.url.pathname
-        } because:  ${(request.response as any)._error.message}`,
+      } because:  ${(request.response as any)._error.message}`,
     })
   }
 })
@@ -260,9 +261,9 @@ server.events.on('request', (request, event) => {
     request.info.remoteAddress
   const sessionId = ip
     ? crypto
-      .createHmac('sha256', secret)
-      .update(ip)
-      .digest('hex')
+        .createHmac('sha256', secret)
+        .update(ip)
+        .digest('hex')
     : 'unknown_ip'
   delete request.headers['x-forwarded-for']
   delete request.headers['x-real-ip']
@@ -274,13 +275,22 @@ async function doWork() {
   try {
     const apolloServer = new ApolloServer({
       typeDefs,
-      resolvers,
+      resolvers: {
+        Query: {
+          vendors,
+          bestPrices,
+          itemsListings: (_, params) =>
+            getScrapeResponses(params, getRedisItems),
+          itemsFlatListings: (_, params) =>
+            getItemsFlatListings(params, getRedisItems),
+        },
+      },
       debug: DEV,
       tracing: DEV,
       cache: !DEV
         ? new RedisCache({
-          host: 'redis',
-        })
+            host: 'redis',
+          })
         : undefined,
       formatError: error => {
         server.log('error', { type: 'graphql-error', error: error.toString() })
