+++ src/api/shared.ts
@@ -2,6 +2,10 @@ import { SOURCES, AMMO_TYPES, RELOAD_TYPES, VENDORS } from '../constants'
 import * as helpers from '../helpers'
 
 import {
+  IItemsFlatListingsOnQueryArguments,
+  IItemFlatListings,
+} from '../graphql-types'
+import {
   IItemsListingsOnQueryArguments,
   IItemListings,
   ItemType,
@@ -9,18 +13,11 @@ import {
   IItemListing,
   SortOrder,
   SortField,
-  IBestPricesOnQueryArguments,
   IBestPrice,
 } from '../graphql-types'
 import { URL } from 'url'
 
-import { getRedisItems } from './redis-getter'
-import { getDyanmoItems } from './dynamo-getter'
-
-
-export async function getBestPrices(
-  params: IBestPricesOnQueryArguments
-): Promise<IBestPrice[]> {
+export async function getBestPrices(): Promise<IBestPrice[]> {
   return Promise.resolve(null)
   // TODO: not used
   // const type = params.type || ItemType.centerfire
@@ -71,7 +68,11 @@ export async function getBestPrices(
 
 export async function getScrapeResponses(
   params: IItemsListingsOnQueryArguments,
-  mode: 'redis' | 'dynamo'
+  valueGetter: (
+    types: ItemType[],
+    subTypes: string[],
+    vendors: string[]
+  ) => Promise<IItemListing[]>
 ): Promise<IItemListings> {
   let {
     itemType,
@@ -104,19 +105,19 @@ export async function getScrapeResponses(
     pageSize = 100
   }
 
-  let types: ItemType[]
+  let itemTypes: ItemType[]
   switch (itemType) {
     case ItemType.reloading:
-      types = RELOAD_TYPES
+      itemTypes = RELOAD_TYPES
       break
     // default to ammo if nothing
     case ItemType.ammo:
     case undefined:
     case null:
-      types = AMMO_TYPES
+      itemTypes = AMMO_TYPES
       break
     default:
-      types = [itemType]
+      itemTypes = [itemType]
   }
 
   // figure out subset of keys to get (keys are stored by vendor)
@@ -130,35 +131,13 @@ export async function getScrapeResponses(
     )
   }
 
-
-  const subTypes = (subType ? [subType] : types.reduce((lst, t) => lst.concat(helpers.itemTypeToStubTypes(t)), []))
-  let results: IItemListing[]
-
-  if (mode === 'redis') {
-    const keys: string[] = types
-      .map(t =>
-        vendors.map(s =>
-          subTypes.map(st =>
-            helpers.getKey(s, t, st)
-          )
-        )
+  const subTypes = subType
+    ? [subType]
+    : itemTypes.reduce(
+        (lst, t) => lst.concat(helpers.itemTypeToStubTypes(t)),
+        []
       )
-      .flat<string>(Infinity)
-    results = await getRedisItems(keys)
-  } else {
-    const keys: string[] = types.reduce(
-      (ll, type) =>
-        ll.concat(
-          subTypes.reduce(
-            (lll, _subType) => lll.concat(`${type}_${_subType}`),
-            [] as string[]
-          )
-        ),
-      [] as string[]
-    )
-
-    results = await getDyanmoItems(vendors, keys)
-  }
+  let results: IItemListing[] = await valueGetter(itemTypes, subTypes, vendors)
 
   const result: IItemListing[] = results
     .filter(
@@ -304,3 +283,119 @@ export async function getScrapeResponses(
     items,
   } as IItemListings
 }
+
+export async function getItemsFlatListings(
+  args: IItemsFlatListingsOnQueryArguments,
+  valueGetter: (
+    types: ItemType[],
+    subTypes: string[],
+    vendors: string[]
+  ) => Promise<IItemListing[]>
+): Promise<IItemFlatListings> {
+  let {
+    itemType,
+    page,
+    pageSize,
+    subType,
+    province,
+    vendor,
+    sortField,
+    sortOrder,
+    query,
+    brand,
+  } = args
+
+  if (!sortOrder) {
+    sortOrder = SortOrder.ASC
+  }
+
+  if (!sortField) {
+    sortField = SortField.minUnitCost
+  }
+
+  if (isNaN(page) || page < 1) {
+    page = 1
+  }
+
+  if (isNaN(pageSize) || pageSize < 1) {
+    pageSize = 25
+  } else if (pageSize > 100) {
+    pageSize = 100
+  }
+
+  let types: ItemType[]
+  switch (itemType) {
+    case ItemType.reloading:
+      types = RELOAD_TYPES
+      break
+    // default to ammo if nothing
+    case ItemType.ammo:
+    case undefined:
+    case null:
+      types = AMMO_TYPES
+      break
+    default:
+      types = [itemType]
+  }
+
+  // figure out subset of keys to get (keys are stored by vendor)
+  let vendors: string[] = SOURCES
+  if (vendor) {
+    const vv = VENDORS.find(v => v.name === vendor)
+    vendors = vv ? [new URL(vv.link).hostname.replace('www.', '')] : []
+  } else if (province) {
+    vendors = VENDORS.filter(v => v.provinces.includes(province)).map(v =>
+      new URL(v.link).hostname.replace('www.', '')
+    )
+  }
+  const subTypes = subType
+    ? [subType]
+    : types.reduce((lst, t) => lst.concat(helpers.itemTypeToStubTypes(t)), [])
+  let results: IItemListing[] = await valueGetter(types, subTypes, vendors)
+
+  // only filters out ammo without subType set (not setup for reloading yet)
+  const result: IItemListing[] = results.filter(
+    r =>
+      r &&
+      r.price > 0 &&
+      (!AMMO_TYPES.includes(itemType) ||
+        (r.subType && r.subType !== 'UNKNOWN')) &&
+      (!brand || (r.brand && r.brand.toLowerCase() === brand.toLowerCase())) &&
+      (!query || r.name.toLowerCase().includes(query.toLowerCase()))
+  )
+
+  // flatten map and sort groups by sortField + sortKey
+  let res = result.sort((a, b) => {
+    let aa = a[sortField]
+    let bb = b[sortField]
+    // put unknown unit costs at the bottom of the sort order
+    if (sortField === SortField.minUnitCost) {
+      if (aa <= 0) {
+        aa = Number.MAX_SAFE_INTEGER
+      }
+      if (bb <= 0) {
+        bb = Number.MAX_SAFE_INTEGER
+      }
+    }
+    const order = sortOrder === SortOrder.ASC ? 1 : -1
+    if (aa === bb) {
+      return 0
+    } else if (aa > bb) {
+      return 1 * order
+    } else {
+      return -1 * order
+    }
+  })
+  // paginate
+  // and then sort vendor listings within the selected groups
+  let items = res.slice((page - 1) * pageSize, page * pageSize)
+
+  return {
+    page,
+    pageSize,
+    pages: Math.ceil(res.length / pageSize),
+    items,
+  } as IItemFlatListings
+}
+
+// todo: merge the 2 functions together...
